Refactor job execution
