+++ src/executor/job/tasks/__init__.py
@@ -1,0 +1,8 @@
+from executor.job.tasks.metadata import update_metadata
+from executor.job.tasks.standard import task_scheduled_job, task_standard_job
+
+__all__ = (
+    "task_scheduled_job",
+    "task_standard_job",
+    "update_metadata",
+)
