+++ src/executor/resource_collector/base.py
@@ -43,4 +43,3 @@ class ResourceIteratorStrategy(ABC):
     ) -> Generator[Resource, None, None]:
         """Iterate over resources and yield Resource models."""
         ...
-
